requirements.txt
datasets==3.1.0
h5py==3.10.0
imageio==2.33.1
librosa==0.10.1
matplotlib==3.7.2
midi2audio==0.1.1
numpy==1.23.5
opencv_python==4.9.0.80
Pillow==11.1.0
PyYAML==6.0.2
scikit_learn==1.2.2
scipy==1.15.2
seaborn==0.13.2
soundfile==0.12.1
svhnl==0.0.5
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.67.0
vector_quantize_pytorch==1.18.0
